+++ TFG/src/main/java/com/alvaroorgaz/easyfactura/model/Factura.java
@@ -12,7 +12,7 @@ public class Factura {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_factura")
-    private Long idFactura;
+    private Integer idFactura;
 
     @ManyToOne
     @JoinColumn(name = "id_empresa", nullable = false)
@@ -33,11 +33,11 @@ public class Factura {
     @OneToMany(mappedBy = "factura")
     private List<FacturaDetalle> detalles;
 
-    public Long getIdFactura() {
+    public Integer getIdFactura() {
         return idFactura;
     }
 
-    public void setIdFactura(Long idFactura) {
+    public void setIdFactura(Integer idFactura) {
         this.idFactura = idFactura;
     }
