+++ TFG/src/main/java/com/alvaroorgaz/easyfactura/service/ClienteService.java
@@ -23,15 +23,15 @@ public class ClienteService {
         return clienteRepository.findAll();
     }
 
-    public List<Cliente> obtenerClientesPorEmpresa(Long idEmpresa) {
+    public List<Cliente> obtenerClientesPorEmpresa(Integer idEmpresa) {
         return clienteRepository.findClientesByEmpresaId(idEmpresa);
     }
 
-    public Cliente obtenerClientePorId(Long id) {
+    public Cliente obtenerClientePorId(Integer id) {
         return clienteRepository.findById(id).orElse(null);
     }
 
-    public void eliminarCliente(Long id) {
+    public void eliminarCliente(Integer id) {
         clienteRepository.deleteById(id);
     }
 }
